# swagger.yaml
swagger: "2.0"
info:
  title: Space Mission Control API
  version: 1.0.1
  description: |
    API for managing space missions, control users, astronauts, and launch vehicles.
    All error strings are required to be clear descriptive message. The descriptive message won't change the marks you receive when being automarked as long as you make an attempt for each one to be custom, but they will affect your code quality mark.
schemes:
  - http
...
components:
  prim:
    ControlUserSessionId:
      type: string
      description: The sessionID representing a logged in controlUser
      example: "abcd1234"
    Error:
      type: string
      example: "Meaningful error message"
    Email:
      type: string
      example: james.kirk@starfleet.gov.au
    Password:
      type: string
      example: livelong123
    NameFirst:
      type: string
      example: John
    NameLast:
      type: string
      example: Carter
...
  group:
    ControlUser:
      type: object
      properties:
        controlUserId:
          type: integer
          example: 132
        name:
          type: string
          example: James Kirk
        email:
          type: string
          example: james.kirk@starfleet.com.au
        numSuccessfulLogins:
          type: integer
          example: 5
        numFailedPasswordsSinceLastLogin:
          type: integer
          example: 2
...
  header:
    ControlUserSessionId:
      in: header
      name: controlUserSessionId
      required: true
      schema:
        $ref: '#/components/prim/ControlUserSessionId'
  body:
...
    ControlUserDetailsUpdate:
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          email:
            $ref: '#/components/prim/Email'
          nameFirst:
            $ref: '#/components/prim/NameFirst'
          nameLast:
            $ref: '#/components/prim/NameLast'
    ControlUserPasswordUpdate:
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          oldPassword:
            $ref: '#/components/prim/Password'
          newPassword:
            $ref: '#/components/prim/Password'
...
  return:
...
    Error:
      properties:
        error:
          $ref: '#/components/prim/Error'
    Empty:
      type: object
    ControlUser:
      type: object
      properties:
        user:
          $ref: '#/components/group/ControlUser'
...
paths:
...
  /v1/admin/controluser/details:
    get:
      summary: Get the details of a control user.
      description: For the given control user that is logged in, return all of the relevant details.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
      responses:
        200:
          description: |
            - "name" is the first and last name concatenated with a single space between them
            - numSuccessfulLogins includes logins direct via registration, and is counted from the moment of registration starting at 1
            - numFailedPasswordsSinceLastLogin is reset every time they have a successful login, and simply counts the number of attempted logins that failed due to incorrect password, only since the last login
          schema:
            $ref: '#/components/return/ControlUser'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
    put:
      summary: Update the details of an control user (non-password).
      description: Given a set of properties, update those properties of this logged in control user.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/body/ControlUserDetailsUpdate'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Email is currently used by another user (excluding the current authorised user)
            - Email does not satisfy this: https://www.npmjs.com/package/validator (validator.isEmail)
            - NameFirst contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
            - NameFirst is less than 2 characters or more than 20 characters
            - NameLast contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
            - NameLast is less than 2 characters or more than 20 characters
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/controluser/password:
    put:
      summary: Update the password of this control user.
      description: Given details relating to a password change, update the password of a logged in user.
      tags:
        # - "By Topic: Access Management"
        - "Iteration 2 (Using Iteration 1)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/body/ControlUserPasswordUpdate'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Old Password is not the correct old password
            - Old Password and New Password match exactly
            - New Password has already been used before by this user
            - New Password is less than 8 characters
            - New Password does not contain at least one number and at least one letter
          schema:
            $ref: '#/components/return/Error'
...